workaround sorting which causing problem by spigot server
according to my debugging, spigot will sent extra package when the
sorting is completed. It seems cause by trying merge the item which is
reach the max staack size, and without synchronize.
this patch just a workaround, it will cause some sort problem when first
sorting. but for me, it better than compete not working at all.