Add an attribute to help build the order of nodes in the composite tree.
The moddeditemtree.xml file contains the reference tree that I used to
make the individual trees.
The initializer only attempts to extract the tree files if
minecraft.tree is missing and the whole system is only used if InvTweaks
config folder exists.
TreeBuilder updated to read and use treeOrder attribute to fine-tune the
insertion of nodes in the composite tree.