Remove modded tree files.
Make the modded tree button go to a repository of modded trees.
Move the merged tree into the system temp folder to minimize the chance
the user would think they could modify it.
Add a readme to the tree folder.
Move all config files to the InvTweaks config folder.
Make a trees folder in the InvTweaks config folder for the tree files.